[REFACTOR] 마이페이지 대시보드 데이터 get api 리팩터링

+++ src/main/java/com/vivid/apiserver/domain/individual_video/application/command/IndividualVideoCommandService.java
@@ -39,7 +39,7 @@ public class IndividualVideoCommandService {
         individualVideo.changeLastAccessTime();
     }
 
-    public void changeProgressRate(IndividualVideo individualVideo, Long progressRate) {
+    public void changeProgressRate(IndividualVideo individualVideo, Integer progressRate) {
         individualVideo.changeProgressRate(progressRate);
     }
